#162: Moved trampolines.
